Versión 1.0.1

+++ src/app/layouts/dashboard/pages/courses/courses.component.ts
@@ -8,31 +8,26 @@ import { LoadingService } from '../../../../core/services/loading.service';
   templateUrl: './courses.component.html',
   styleUrl: './courses.component.sass'
 })
+
 export class CoursesComponent {
-  displayedColumns: string[] =  ['id', 'coursesName', 'startDate', 'endDate', 'actions' ]
+  displayedColumns: string[] = ['id', 'coursesName', 'startDate', 'endDate', 'actions']
   courses: Courses[] = []
 
   constructor(private coursesService: CoursesService, private loadingService: LoadingService) {
-    this.loadingService.setIsLoading(true);
     this.coursesService.getCourses().subscribe({
-        next: (courses) => {
-          this.courses = courses;
-        },
-        complete:  () => {
-          this.loadingService.setIsLoading(false);
-        },
+      next: (courses) => {
+        this.courses = courses;
+      },
     });
   }
 
   onDelete(id: number) {
-    this.loadingService.setIsLoading(true);
-    this.coursesService.deleteCourseById(id).subscribe({
+    if (confirm('¿Desea eliminar la selección?')) {
+      this.coursesService.deleteCourseById(id).subscribe({
         next: (courses) => {
           this.courses = courses;
         },
-        complete:  () => {
-          this.loadingService.setIsLoading(false);
-        },
-    })
+      })
+    }
   }
 }
